refactor(mobile): extract entry list article component
